+++ api/holidays/holidays.go
@@ -3,26 +3,22 @@ package holidays
 import (
 	"encoding/json"
 	"main/api"
+	"main/db"
 	"main/debug"
 	"net/http"
 	"strings"
+	"time"
 )
 
 // Struct for information about one holiday, used when getting data from the API
 type Holiday struct {
 	Date string `json:"date"`
-	LocalName string `json:"localName"`
 	Name string `json:"name"`
-	CountryCode string `json:"countryCode"`
-	Fixed bool `json:"fixed"`
-	Global bool `json:"global"`
-	Counties []string `json:"counties"`
-	LaunchYear int `json:"launchYear"`
-	Type string `json:"string"`
 }
 
-func GetCountryHolidays(w http.ResponseWriter, r* http.Request) {
+func GetCountryHolidays(w http.ResponseWriter, r *http.Request) {
 	var countryHolidays []Holiday
+	var newCountryHolidays = make(map[string]interface{}, len(countryHolidays))
 
 	// Parsing variables from URL path
 	path := strings.Split(r.URL.Path, "/")
@@ -37,26 +33,74 @@ func GetCountryHolidays(w http.ResponseWriter, r* http.Request) {
 		debug.ErrorMessage.Print(w)
 		return
 	}
+
 	year := path[4]
 	location := path[5]
 
-	// Get the country's holiday
-	countryHolidays, status, err := GetAllHolidays(year, location)
-	if err != nil {
+	// Check if country is already stored in the database
+	data, exist, err := db.DB.Get("Holidays", location)
+	if err != nil && exist {
 		debug.ErrorMessage.Update(
+			http.StatusInternalServerError,
+			"WeatherData.Handler() -> Database.get() -> Trying to get data",
+			err.Error(),
+			"Unknown",
+		)
+		debug.ErrorMessage.Print(w)
+		return
+	}
+
+	if exist {
+		// Convert the data to a map
+		m := data.Container.(map[string]interface{})
+
+		// Assign the values to the output map
+		for key, elem := range m {
+			newCountryHolidays[key] = elem
+		}
+	} else {
+		// Get data from the API and add to the database
+		var status int
+		countryHolidays, status, err = GetAllHolidays(year, location)
+		if err != nil {
+			debug.ErrorMessage.Update(
 				status,
 				"holidays.GetCountryHolidays() -> holidays.GetAllHolidays() -> Getting holidays",
 				err.Error(),
 				"Unknown",
 			)
-		debug.ErrorMessage.Print(w)
-		return
+			debug.ErrorMessage.Print(w)
+			return
+		}
+
+		// Put the holidays data in a map where the key is the name and value is the date
+
+		for i := 0; i < len(countryHolidays); i++ {
+			newCountryHolidays[countryHolidays[i].Name] = countryHolidays[i].Date
+		}
+
+		// Add data to the database
+		var data db.Data
+		data.Time = time.Now().String()
+		data.Container = newCountryHolidays
+
+		_, err = db.DB.Add("Holidays", location, data)
+		if err != nil {
+			debug.ErrorMessage.Update(
+				http.StatusInternalServerError,
+				"holidays.GetCountryHolidays() -> Database.Add() -> Adding data to database",
+				err.Error(),
+				"Unknown",
+			)
+			debug.ErrorMessage.Print(w)
+			return
+		}
 	}
 
 	// Sending the response to the user
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
-	err = json.NewEncoder(w).Encode(countryHolidays)
+	err = json.NewEncoder(w).Encode(newCountryHolidays)
 	if err != nil {
 		debug.ErrorMessage.Update(
 			http.StatusInternalServerError,
@@ -65,12 +109,11 @@ func GetCountryHolidays(w http.ResponseWriter, r* http.Request) {
 			"Unknown",
 		)
 		debug.ErrorMessage.Print(w)
-
 	}
 }
 
 // Get information about all holidays in a country
-func GetAllHolidays(year string, countryCode string) ([]Holiday, int,  error) {
+func GetAllHolidays(year string, countryCode string) ([]Holiday, int, error) {
 	// Slice with holiday structs
 	var countryHolidays []Holiday
 	url := "https://date.nager.at/api/v2/PublicHolidays/" + year + "/" + countryCode
@@ -89,4 +132,3 @@ func GetAllHolidays(year string, countryCode string) ([]Holiday, int,  error) {
 
 	return countryHolidays, http.StatusOK, err
 }
-